requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gspread>=6.0.0
google-auth>=2.27.0
pandas>=2.0.0
scipy>=1.11.0
numpy>=1.24.0
python-dotenv>=1.0.0
cachetools>=5.3.0
apscheduler>=3.10.0
pytz>=2024.1
